pwkit/cli/imtool.py: bugfix for multi-plane images on newer numpy

+++ pwkit/cli/imtool.py
@@ -53,7 +53,7 @@ WCS support isn't fantastic and sometimes causes crashes."""
         except Exception as e:
             print ('blink: can\'t convert “%s” to simple 2D sky image; taking '
                    'first plane' % path, file=sys.stderr)
-            data = img.read (flip=True)[tuple (np.zeros (img.shape.size - 2))]
+            data = img.read (flip=True)[tuple (np.zeros (img.shape.size - 2, dtype=np.int))]
             toworld = None
         else:
             data = img.read (flip=True)
@@ -503,7 +503,7 @@ WCS support isn't fantastic and sometimes causes crashes."""
             except Exception as e:
                 print ('imtool show: can\'t convert “%s” to simple 2D sky image; taking '
                        ' first plane' % path, file=sys.stderr)
-                data = img.read (flip=True)[tuple (np.zeros (img.shape.size - 2))]
+                data = img.read (flip=True)[tuple(np.zeros (img.shape.size - 2, dtype=np.int))]
                 toworld = None
             else:
                 data = img.read (flip=True)
